Adding support for multiple networks
This commits adds changes to assign an interface name
per network configuration. The runtime can set the default
network interface name it is interested in and can access the
network information based on the default interface. CNI
spec doesnt prescribe a convention for a default network.
So convention used in the library will be to use the first
network config in the sorted network config set as the default
network and assiged the default interface to that network.
Subsequent networks created will be assigned a generated
network interface.

Signed-off-by: abhi <abhi@docker.com>